youtube_transcript_adk/requirements.txt
youtube-transcript-api>=0.5.0
google-adk
python-dotenv>=1.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
